Add method 'AddYChannel' to TimePlot
The structure of the class TimePlot was generalized for plotting multiple channels at once.

Method 'AddYChannel' allows the addition of a curve in a TimePlot.

Add new class 'DataListener' to hold the data received in a ychannel, so each one can be monitored individually.

For each ychannel, method AddYChannel creates a line in data_buffer, a DataListener and a curve (with a given color).

If TimePlot has more than one ychannel, 'updatesAsynchronously' is automatically set to True, so all data can be draw correctly.

Time array is now stored in index 0 of data_buffer. Indexes higher than 0 are for ychannel arrays.