Plot widget improvements (#60)
Adds support for multiple curves in PyDMTimePlot and PyDMWaveformPlot. In designer, you can add or remove curves from a plot by right clicking it and choosing 'Edit Curves', or by double-clicking the plot. This will bring up a dialog that lets you edit the curves. All curves info is stored in a QStringList property called 'curves', each item in the list is a JSON dictionary with properties for the curve.

* Add method 'AddYChannel' to TimePlot

The structure of the class TimePlot was generalized for plotting multiple channels at once.

Method 'AddYChannel' allows the addition of a curve in a TimePlot.

Add new class 'DataListener' to hold the data received in a ychannel, so each one can be monitored individually.

For each ychannel, method AddYChannel creates a line in data_buffer, a DataListener and a curve (with a given color).

If TimePlot has more than one ychannel, 'updatesAsynchronously' is automatically set to True, so all data can be draw correctly.

Time array is now stored in index 0 of data_buffer. Indexes higher than 0 are for ychannel arrays.

* Move method connectionStateChanged to class DataListener

The state of connection of a ychannel is now monitored inside its DataListener. If a channel is not connected, its curve is not plotted (to prevent errors/warnings and misleading info for the user).

* Add Example: TimePlot with Multiple Curves

* Clean comments in timeplot_multiple_curves example

* Add 'showLegend' property to BasePlot

Method addYChannel in class TimePlot now adds an item to the legend if a showLegend is True.

* Check if ychannel is empty string before calling addYChannel

* Add self.curve to the list only if self._ychannel is defined

This prevents unnecessary curves in the list.

* Verify if parameter ychannel is defined before creating new channel

* Add new utilities for making colors a little more user friendly.

* Move multiple curve functionality into base plot, revise time plot and waveform plots to utilize the new capabilities.

* Add an option to return hex string if there is no corresponding svg color.

* WIP - Custom Plugin for Time Plots with edit menu actions.  Doesn't work yet.

* Add Designer Task Menu extension for PyDMTimePlot.

* Don't clear existing text when editing a curve field in the table.

* Ensure color string is always a python string.

* Remove commented-out, unused sort method.

* Improved comments.

* Get rid of unused import.

* Disable the time plot redraw timer if in Designer.

* Add curve edit menu to the waveform plot.

* Don't waste space showing a blank title.

* Update scope example to use new waveform plot.

* Use waveform slot instead of value slot.

* Fix swapped X and Y channels in dict representation.

* Set better defaults for curve editor sizes.

* Fix several issues with handling empty fields in curve editors.

* Clean up plotting examples.

* Use PyQt compatibility layer, don't import directly from PyQt4.

* Treat pickle files as binary files when opening.

* Add QItemSelection to both QtCore and QtGui, so that people used to both PyQt4 and PyQt5 can import it from where they expect.

* Remove check on init_x_channels that always evaluated to true.

* Fix problem with not using relative imports.

* Fix waveform plot editor import issues.

* Open pickled color files as binaries.

* Fix python3 compatibility issues with handling QVariants.

* Refactor WaveformPlotCurveItem initializer.

* Remove unused import statement.