PR improvements in plot classes (#160)
Refactor plotting code, now all plot curve items have a common base class.
* Bug Fix in scatter_plot:

The code was using the NoDataError class from waveformplot but was not
importing it. The solution envolves defining the class in baseplot.py
and importing it in both widgets.

* Create BasePlotCurveItem to serve as base class for all curves in PyDM.

* Adapt ScatterPlotCurveItem to derive from BasePlotCurveItem.

* Adapt TimePlotCurveItem to derive from BasePlotCurveItem.

* Adapt WaveformCurveItem to derive from BasePlotCurveItem.

* Add redraw types to BasePlotCurveItem class.

* Add bufferSize to curves properties in PyDMScatterPlot class.

* Import cleanup and bug fix in initialization of PyDMScatterPlot.

* Finish adaptation of PyDMTimePlot and PyDMTimePlotCuveItem.

* Change in QtDesigner classes of plot widgets.

Now there is Base classes for all QtDesigner classes.