TST: Eliminate race condition at rules test.
time.sleep was blocking the qapp to process events and with that the rules were never being executed unless it happened between the sleep and the next block of code.
this caused the tests to fail sometimes. using qtbot.wait makes the app still be responsive while we wait for a certain period of time to ensure that the rule was executed.