ENH: provide a decorator to add rule properties to a class
The current way in which rule-triggered properties are set for `PyDMPrimitiveWidget` derivative classes is a bit cumbersome in that:

* It's done at `__init__` time.
* Requires copying the rules from the parent class, which creates unneeded coupling.
* Forces us to write boilerplate code just to check if it's the first time this is being done, and only then add the extra properties.

The first point is not a big issue in practice because Designer instantiates the classes after loading them, but not ideal (and we're doing class-related stuff from an instance...)

The second point is not a huge problem either because you don't refactor parent class names that often, but not pretty either.

The third point is more salient. Boilerplate code is hard to refactor (if needed), and is prone to bugs.

In this commit I introduce an `__init_subclass__` method for `PyDMPrimitiveWidget` which allows updating the derivative `RULE_PROPERTIES` just by providing a new dictionary augmenting the inherited one.