FIX: Line edit value setting on Python 3.10+
In this method, when 'channeltype' is 'int', we correctly set both the
value and 'scale' to integers. However, we didn't set the resulting
division to an integer as well. Since integer division can result in a
non-integer number, the result is sometimes (maybe all the times?) a
'float' type.

This didn't cause much trouble before Python 3.10, since Python would
automatically convert the floating-point number to an integer. At most
you got a Deprecation warning telling you that this isn't the best thing
to do. However, starting with 3.10, that conversion no longer works, so
the resulting number that gets sent to the communication channel is a
mess no longer resembling the original, causing issues.

This patch simply wraps the division into an explicit conversion to
'channeltype', avoiding the issue.

Signed-off-by: flowln <thiago.ferreira@lnls.br>