ENH: Allow following symlinks when opening files via rel. paths
This adds a 'followSymlinks' property to both PyDMEmbeddedDisplay and
PyDMRelatedDisplayButton, so that the searching behavior in them can be
modified to canonize the base path before looking for the pointed-to
display file.

This improves our ability to have PyDM interfaces installed via the
usage of symbolic links in deployment machines.

The default behavior to not follow those symlinks is retained for
backwards compatibility.

Signed-off-by: flowln <thiago.ferreira@lnls.br>