BUG: fix issue where designer doesn't load pydm widgets when pydm is installed from PyPI using pip
When pydm is installed using pip, you should be able to
set the env variable PYQTDESIGNERPATH to the pip pydm install location and then designer loads the pydm widgets.

But for PYQTDESIGNERPATH to be used to load pydm widgets, it must point to dir containing
pydm_designer_plugin.py file. This file was wrongly placed in the pydm repo
root dir. But pip only copies to the install location any python-package
dirs (dir with a __init__.py), which for pydm is the /pydm subfolder
of the root dir (which did not contain the pydm_designer_plugin.py file).

So solution is to place the plugin file in the /pydm subdir so pip
copies it to the install location, and so PYQTDESIGNERPATH can find the
file when set to the install location.

(conda installs are not affected, does not use pydm_designer_plugin.py
file)