MNT: remove widget classes inheriting from enum, as preperation for qt6 enum support
The need for inheritance was b/c  when an enum gets written to a ui file by
designer, its in the form of <widget class name>::<enum value> (ex='PyDMLabel::STRING').

On loading the ui file, unless the widget class had subclassed the enum class it
would not be able to find the definition of the enum.

The issue with subclassing the enum is:  in qt6 enums must inherit
from python Enum (ex = class MyNewEnum(Enum)), and these classes can not be
inherited from.

So instead of inheriting the enum we can just re-declare the values of the enum
in the class itself.