MNT: Deal with pyside6 throwing an error about __init__ functions in classes with multiple inheritence not getting called, when we actually are calling them.
There is no issue running on pyqt5, but on pyside6 throws a runtime error when initializing classes
that use multiple inheritance of a qt base class and a pydm class.

example class:
    class PyDMDrawing(QWidget, PyDMWidget, new_properties=_penRuleProperties))
    ...
    def __init__(self, parent=None, init_channel=None):
       ...
       # explicitly calling base class __init__ functions
       QWidget.__init__(self, parent)
       PyDMWidget.__init__(self, init_channel=init_channel)
       ...

The runtime error happens when a call is made in the base class that
relies on the qt base class's __init__ call to have completed setting
things up (such as 'installEventFilter()).

example error:
    -> self.installEventFilter(self)
    RuntimeError: '__init__' method of object's base class (PyDMDrawingImage) not called.

This error is thrown despite that we explicitly call each of it's base class __init__
functions. It seems like pyside6 is just being extra cautious here.

To avoid a larger rework of the widget class structures just to appease
this pyside6 error, I propose we just catch this error where it occurs.