MNT: fix segfault issue when using symbols on pyside6
The symbol classes can intentionally try to read in a .jpg that does
not exit, such as in the symbol editor tests.

Note from original symbol editor test commit:
"tests reference a file "goodbye.jpg" as a reference image to test with, no such file exists but not a problem––to increase coverage, add a file "goodbye.jpg" (can be any jpg image) to the same directory"

But executing the file-reading calls with a non-existent file segfaults on pyside6, so make sure they are not called in this case.