MNT: update enum-button widget's button-group to use proper clicked signal, avoid pyside6 error
Calling signal without overloaded int and expecting a button object back,
is whats specified in qt5 docs:
https://doc.qt.io/qt-5/qbuttongroup.html#buttonClicked.

This is the same in the qt6 docs, and pyside6 even throws error if signal is not
used in this correct way.

Also seems like the newer python in running pyside6 on doesn't like using enums
to index into an array, so use a map for 'class_for_type' instead.
(using the enum's '.value' should allow for indexing into an array, but imo using the map is more readable.)