MNT: switch away from python2 style super calls to python3
The background on this patch is that running on pyside6 exposes some weirdness which how we are doing super() __init__ calls, so this it comes as some cleanup before another fix that will follow.

python2 style: super(ClassName, self).__init__()
python3 style: super().__init__()

While these python2 style calls work on python3, the new way is cleaner and imo it's best to follow python3's conventions as we officially don't support python2.

references:
https://docs.python.org/3/library/functions.html#super
https://stackoverflow.com/a/19776143