MNT: change widget class inheritence ordering to change it's 'mro' (method resolution order) and appease pyside6.
In python, the 'mro' of a python class is determined by the ordering in
its definition of the classes it inherits from, for ex:

class PyDMByteIndicator(PyDMWidget, QWidget):
->  Method resolution order:
 |      PyDMByteIndicator
 |      pydm.widgets.base.PyDMWidget
 |      pydm.widgets.base.PyDMPrimitiveWidget
 |      PySide6.QtWidgets.QWidget
 |      PySide6.QtCore.QObject
 |      PySide6.QtGui.QPaintDevice
 |      Shiboken.Object
 |      builtins.object

This ordering will effect which parent class __init__ call will be done
first when doing 'super().__init__()'

This ordering effects some other things too, such as if both parent
classes have a same named function it decides which gets executed.

When using PySide6, we get thrown an error if the mro of classes that
inherit from both a PyDM widget and a Qt widget does not have the Qt
widget resolving fist, so for ex we must have:

class PyDMByteIndicator(QWidget, PyDMWidget): # swap ordering of inherited classes
->  Method resolution order:
 |      PyDMByteIndicator
 |      PyQt5.QtWidgets.QWidget # Qt class initialized first
 |      PyQt5.QtCore.QObject
 |      sip.wrapper
 |      PyQt5.QtGui.QPaintDevice
 |      sip.simplewrapper
 |      pydm.widgets.base.PyDMWidget # PyDM class initialized second
 |      pydm.widgets.base.PyDMPrimitiveWidget
 |      builtins.object

This is b/c our PyDM widget __init__ methods make calls that require
it's Qt parent-class to have been initialized first. (ex:
self.installEventFilter(self)). And while we do explicitly call the
parent class __init__ functions in the correct ordering, for ex:

    def __init__(self, parent: Optional[QWidget] = None, init_channel=None):
        QWidget.__init__(self, parent)
        PyDMWidget.__init__(self, init_channel=init_channel)

this doesn't seem to convince PySide6 that the initialization will
happen in the correct ordering. Only the mro ordering being correct
appeases PySide6 and stops the error.

Note: we can't use 'super().__init__()' here and rely solely on the mro
to resolve the ordering, for example in this case b/c QWidget __init__
takes a positional arg and PyDMWidget's takes a different named arg.