BUG: fix error when establishing plugin connections with protocols in all caps
We only defined and only accept the protocol in channel addresses when its all lowercase. ('pva', 'calc', 'ca', etc)
Used in a while address, it looks like: 'ca://test:example'.

But sometimes we set the protocol to be uppercase, like in the
testing when we use 'CA://MTEST'.

This fix just converts all protocols into lowercase to avoid an error,
which imo is more user friendly than enforcing case-sensitive protocols

Found this issue when running tests on pyside6 since with it logger.error() calls are printed and fail
the tests, but on pyqt5 we need to set 'log_cli=true' to get the same
behavior. (will set this in followup patch)