MNT: on pyside6, ensure events get filtered properly into pydm's base classes
Some more pyside6 weirdness was that events such as middle-click's
were not getting propated from child classes of PyDMWidget/PyDMWritableWidget
to their parent's eventFilter() funcs.
This is even after an earlier fix for event-handling, which made pyqt5's events work again but not
pyside6's.

The problem seems to be that pyside6 is unsure of which parent class to
propagate the events to, and explicity declaring this in the
eventFilter() func of the child classes fixes the issue.