BUG: stop designer from setting line-edits in ui file to be readOnly when PYDM_DESIGNER_ONLINE enabled
when editing a lineedit widget in designer after doing 'export PYDM_DESIGNER_ONLINE=1',
we want widgets to act as read-only so PVs are not accidentally changed
when modifying screens. this is enforced by the dataplugins themselves
and is_read_only() (set during designer setup: https://github.com/slaclab/pydm/blob/master/pydm/qtdesigner.py#L47)

but we don't want the readOnly=true status to be saved into the
ui file containing the lineedit.

to avoid this, don't let lineedit set itself to readOnly if it's
loaded in designer with live data enabled. (it will still act in a
read-only way thanks to dataplugins themselves)