BUG: make tabwidget call parent eventFilter() of correct parent class.
Stops the following error output when running a PyDMTabWidget widget:

Traceback (most recent call last):
  File "/home/nolan/repos/pydm/pydm/widgets/tab_bar.py", line 36, in eventFilter
    return PyDMWritableWidget.eventFilter(self, obj, event)
NameError: name 'PyDMWritableWidget' is not defined. Did you mean: 'PyDMTabWidget'?

We do these explicit parent eventFilter() calls in the first place so
events to get events to be correctly on pyside6.

This bug got merged in b/c of lacking tests for the PyDMTabWidget, but
tests for it and PyDMTerminator and PyDMNTTable will be added shortly.