ENH/MNT: convert all Qt `Property` declarations to use a way that works for both pyside6 and pyqt5.
the current way we use the `@Property` function-decorator doesn't work
with pyside6.

the docs for how to declare Properties in pyside6 are
here: https://doc.qt.io/qtforpython-6/PySide6/QtCore/Property.html.

the docs claim that `@Property` should work if you import it from
`PySide6.QtCore`, but from my testing it does not.

what does work is the `Property()` function, which has an equivalent
in pyqt5 `pyqtProperty()`. So the idea of this patch is to switch all
instances of `@Property` to their equivalent function calls, and
conditionally import them depending on the current API.